+++ docs/changelog.md
@@ -10,6 +10,8 @@
   - 新增 `Festival.description` 属性，表示规范化中文描述
 - `borax.calendars.utils`
   - 新增 `ThreeNineUtils` 三伏九九天解析工具
+- 开发SOP
+  - 移除whl安装包文件的py2标识 
 
 ## v3.5.0 (20211115)